smartresearch/app/generator.py
"""
Generator — sends retrieved context + user question to Google Gemini to produce an answer.

Uses the new 'google-genai' SDK (replacement for the deprecated 'google-generativeai').
"""
from typing import List, Dict, Any

from google import genai
from google.genai import types

from app.config import GEMINI_API_KEY, GEMINI_MODEL


SYSTEM_PROMPT = """\
You are SmartResearch, an AI research assistant.
Answer the user's question using ONLY the provided context chunks.
If the context does not contain enough information, say so honestly.
Always cite the source document and page number at the end of your answer.
Be concise, accurate, and helpful.
"""


def build_context(chunks: List[Dict[str, Any]]) -> str:
    """Format retrieved chunks into a readable context block."""
    parts = []
    for i, chunk in enumerate(chunks, 1):
        parts.append(
            f"[Chunk {i} | Source: {chunk['source']} | Page: {chunk['page']}]\n{chunk['text']}"
        )
    return "\n\n---\n\n".join(parts)


def generate_answer(query: str, chunks: List[Dict[str, Any]]) -> str:
    """
    Generate a grounded answer using Gemini given the query and top-K retrieved chunks.

    Args:
        query:  The user's natural language question.
        chunks: Retrieved context chunks from Endee.

    Returns:
        The generated answer as a string.
    """
    if not GEMINI_API_KEY:
        raise EnvironmentError(
            "GEMINI_API_KEY is not set. Please copy .env.example to .env and add your key."
        )

    context = build_context(chunks)
    prompt = (
        f"{SYSTEM_PROMPT}\n\n"
        f"=== CONTEXT ===\n{context}\n\n"
        f"=== QUESTION ===\n{query}\n\n"
        f"=== ANSWER ==="
    )

    # New google-genai SDK
    client = genai.Client(api_key=GEMINI_API_KEY)
    try:
        response = client.models.generate_content(
            model=GEMINI_MODEL,
            contents=prompt,
        )
        return response.text.strip()
    except genai.errors.APIError as e:
        if e.code == 429:
            return "⚠️ Gemini API free-tier rate limit exceeded (15 requests/minute). Please wait 45 seconds and try again!"
        return f"⚠️ Gemini API Error: {str(e)}"
    except Exception as e:
        return f"⚠️ Error generating answer: {str(e)}"
