smartresearch/requirements.txt
# Vector database client
endee>=0.1.0

# Embeddings (local, no API key needed)
sentence-transformers>=2.7.0

# LLM (Google Gemini)
google-genai>=1.0.0

# Document loaders
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Text chunking
langchain-text-splitters>=0.2.0

# Environment variable management
python-dotenv>=1.0.0

# Web UI
streamlit>=1.35.0

# CLI
click>=8.1.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
